Benchmark data collection completed.
For the pre warm collection we use the prewarm processor activation time only, since that's what matters the most. The xgb model's prediction for prewarm decides if we create prewarm processors at all. If it's 1 then we create but then base don the probability. Based on the probability we create up to 3 prewarm processors and we log the pre warm activation time on every scheduling run.